requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
Flask==2.0.3
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.1
Jinja2==3.1.0
MarkupSafe==2.1.1
numpy==1.22.3
oauthlib==3.2.0
pandas==1.4.1
psycopg2==2.9.3
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
SQLAlchemy==1.4.33
tweepy==4.8.0
urllib3==1.26.9
Werkzeug==2.0.3
